Use floats for all instruments

+++ reporter.go
@@ -7,7 +7,7 @@ type Reporter interface {
 	// allows reporters to prepare for next data snapshot.
 	Prep() error
 	// Discrete accepts a numeric value with name and (sorted) tags
-	Discrete(name string, tags []string, value int64) error
+	Discrete(name string, tags []string, value float64) error
 	// Sample accepts a sampled distribution with name and (sorted) tags
 	Sample(name string, tags []string, dist Distribution) error
 	// Flush is called at the end of each reporting cycle, which
